Extract stateless conversation list primitives into common package
