+++ app/src/test/kotlin/com/android/messaging/ui/conversationlist/ui/ConversationListContentTest.kt
@@ -1,7 +1,8 @@
 package com.android.messaging.ui.conversationlist.ui
 
+import com.android.messaging.ui.conversationlist.common.previewConversationListItem
 import com.android.messaging.ui.conversationlist.model.ConversationListItemUiModel
-import com.android.messaging.ui.conversationlist.ui.support.previewConversationListItem
+import com.android.messaging.ui.conversationlist.common.previewConversationListItem
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNull
 import org.junit.Test
