+++ src/com/android/messaging/ui/conversationlist/common/ConversationListItemAvatar.kt
@@ -1,4 +1,4 @@
-package com.android.messaging.ui.conversationlist.ui.item
+package com.android.messaging.ui.conversationlist.common
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
